Inheritence.net
Public property of parent class is accessible in child class
 due to inheritance. 
 t1.showDim();
Does class Triangle have showDim() method? ?
No. so if it is not fount in class Triangle it will go to its 
parent class and execute that method


Triangle t1 = new Triangle() { width=4.0,height=4.0,style ="isosceles"};
With this line noargs constructor of parent and child 
get called