Create Publisher_subscriber_design&event .net
For eg.Vodaphone may announce some scheme for all user who have the plan ABC. The announcement will be issued to all subscriber of the plan ABC.
Here it has team who’s job is to just announce. Whom it will be received they are not aware of. The team says my job is to announce.
Now All Vodaphone user with plan ABC will automatically get the message of scheme. This is because all theses user have subscribe to that service. 

So here following code says if SomeEvent is pointing to method then call it. if (SomeEvent != null) this line ensure that at least  one person has subscribed to method(service) other wise it should not call it. Ie if no one has subscribe to (method)service then how can it call that method.