FullAyonymous.cs
              Pass Arguments to an Anonymous Method

It is possible to pass one or more arguments to an anonymous method. 
To do so, follow the delegate keyword with a parenthesized parameter list. Then, pass the argument(s) to the delegate instance when it is called.
In this version, CountIt now takes an integer argument. Notice how the parameter list is specified after the delegate keyword when the anonymous method is created. The code inside the anonymous method has access to the parameter end in just the same way it would if a named method were being created. The output from this program is shown next:
0
1
2
3
0
1
2
3
4
5